move widgets to widgets module
