+++ leetcode-tui-rs/src/app.rs
@@ -1,5 +1,5 @@
 use crate::utils::update_database_questions;
-use crate::{ctx::Ctx, executor::Executor, root::Root, signals::Signals};
+use crate::{ctx::Ctx, executor::Executor, signals::Signals, widgets::root::Root};
 use color_eyre::Result;
 use leetcode_tui_config::{constants::EDITOR, key::Key};
 use leetcode_tui_core::{emit, Event, UBStrSender};
