+++ leetcode-tui-rs/src/lib.rs
@@ -1,11 +1,6 @@
 pub mod app;
 pub mod ctx;
 pub mod executor;
-pub mod help;
-pub mod popup;
-pub mod progress_bar;
-pub mod question;
-pub mod root;
 pub mod signals;
-pub mod topic;
 pub mod utils;
+pub mod widgets;
